requirements.txt
flask
bcrypt
cryptography
sqlalchemy
python-dotenv
requests
